+++ cs/TagsCloudVisualization/ArchimedeanSpiral.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Drawing;
 
-namespace ProjectCircularCloudLayouter
+namespace TagsCloudVisualization
 {
     public class ArchimedeanSpiral : ISpiral
     {
