+++ cs/TagsCloudVisualization/CircularCloudLayouter.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 
-namespace ProjectCircularCloudLayouter
+namespace TagsCloudVisualization
 {
     public class CircularCloudLayouter
     {
