+++ cs/TagsCloudVisualization/CircularCloudVisualisation.cs
@@ -2,7 +2,7 @@
 using System.Drawing;
 using System.Drawing.Imaging;
 
-namespace ProjectCircularCloudLayouter
+namespace TagsCloudVisualization
 {
     public static class CircularCloudVisualisation
     {
