+++ cs/TagsCloudVisualization/GenerateRectangles.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Drawing;
 
-namespace ProjectCircularCloudLayouter
+namespace TagsCloudVisualization
 {
     public static class GenerateRectangles
     {
