+++ cs/TagsCloudVisualization/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System.Drawing;
 using System.Drawing.Imaging;
 
-namespace ProjectCircularCloudLayouter
+namespace TagsCloudVisualization
 {
     class Program
     {
