+++ cs/TagsCloudVisualizationTests/TagsCloudVisualizationTests.csproj
@@ -14,7 +14,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\ProjectCircularCloudLayouter\ProjectCircularCloudLayouter.csproj" />
+      <ProjectReference Include="..\TagsCloudVisualization\TagsCloudVisualization.csproj" />
     </ItemGroup>
 
 </Project>
